runDocker.BAT for building and running docker from local files + message about missing configuration.
runDocker.BAT will build the docker image from the local files (not from remote repository)
and use the local files in the docker image which makes updating and testing them easier.

Updated 'measure.py' to hint that the .env file needs to be created.

The README.md contains setup instruction that require to execute python commands
that did not seem to apply to the docker setup.
But in fact, setting up '.env' still applies.

The README.md could also be updated, but I added an exit message to
measure.py.

I do not know if this is the "docker" way of doing things,
but this was an occasion for me to have another look at docker -;).