Change values for measurement
Shelly is kinda slow to respond with new values to changes of the bulb. That's why it makes sense to have a generous sleep time between individual measurements, to avoid reading the "previous value".

Since the overall sleep time is pretty generous we can reduce the sleep time for sat/hue/ct changes.

---


I further reduced the steps to the default of HA:

24 different colors in a ring with 8 saturation steps. Also I think 5 brightness steps allows fair estimations if between them.

This results in a measuring time of around 12 hours instead of 36 which I previous got with longer sleep intervals (because of shellys slow response.

--- 

I increase the timeout for the shelly. There's nothing wrong with waiting a bit longer instead of aborting - since we talk about wifi here.