Use the new smart_dimmer class for the HmIP-BDT device
The device type was `light` before, which meant that no fixed power
could be specified by the user in the UI config flow; see #2693 for
discussion.

In response, a new device type `smart_dimmer` was created in #2765.

[This comment](https://github.com/bramstroker/homeassistant-powercalc/issues/2754#issuecomment-2524028107)
describes how to use the new device type; the changes in this commit
follow that pattern.